fix deleting applications
Because in C you can only delete a directory if nothing is in it.
and you can't unlink a directory, the current seemingly odd implementation
of a manual rmdir -r function is actually correct - except it hasn't stood
the test of time.

You can only unlink files - you cannot unlink a directory. So in the function
unlink will be used on files, and rmdir will be used once it's all set and done.